Moved classes in the subsystem package
